requirements.txt
asgiref==3.5.2
attrs==22.1.0
backports.zoneinfo==0.2.1
Django==4.1.2
djangorestframework==3.14.0
drf-spectacular==0.24.2
importlib-resources==5.10.0
inflection==0.5.1
jsonschema==4.16.0
pkgutil-resolve-name==1.3.10
pyrsistent==0.18.1
pytz==2022.5
PyYAML==6.0
sqlparse==0.4.3
uritemplate==4.1.1
zipp==3.10.0
